- Remove custom Enum code in CommandService and extend existing Enum code to use ChoiceDisplay if it exists - Move Formula and Craft command to their own file - Updated Formula and Craft commands to be ephemeral initially and then respond with a new message
