+++ EGG9000.Bot/Commands/MiscCommandsSlash.cs
@@ -31,96 +31,6 @@ using static Ei.Backup.Types;
 
 namespace EGG9000.Bot.Commands {
     public static class MiscCommandsSlash {
-        [SlashCommand(Description = "Show you required artifacts to craft the requested artifact.")]
-        public static async Task Craft(FauxCommand command, [SlashParam(Description = "Quantity")] int quantity, [SlashParam]TierInput quality, [SlashParam(AutocompleteHandler = typeof(EggIncArtifacts.ArtifactNameAutoComplete))] string artifact, ApplicationDbContext db, ILogger logger) {
-            var requestedArtifact = EggIncArtifacts.GetEiAfxData().artifact_families.FirstOrDefault(x => x.id == artifact);
-
-            if(requestedArtifact is null) {
-                await command.RespondAsync($"Unable to locate an artifact with the name {artifact}");
-                return;
-            }
-
-            await command.RespondAsync("Getting backups...");
-
-            var user = await db.DBUsers.FirstOrDefaultAsync(x => x.DiscordId == command.User.Id);
-            if(user == null) {
-                await command.RespondAsync("⚠️ERROR: Unable to find backups for this user");
-                return;
-            }
-
-            var contentString = "";
-
-            if(user.EggIncAccounts.Count == 1) {
-                contentString = await CraftStringBuilder(user.EggIncAccounts.First(), quantity, quality, requestedArtifact);
-                await command.ModifyOriginalResponseAsync(x => {
-                    x.Embed = null;
-                    x.Content = $"\n{contentString}\n";
-                });
-            } else {
-                var builder = new ComponentBuilder();
-                foreach(var account in user.EggIncAccounts) {
-                    builder.WithButton($"{account.Name} {account.Backup?.EarningsBonus.ToEggString()}", customId: $"CraftAccountButton:{account.Id}|{((int)quality)}|{quantity}|{artifact}");
-                }
-                await command.ModifyOriginalResponseAsync(x => { x.Content = "Please select the account you would like to craft with."; x.Components = builder.Build(); });
-            }
-
-            user.UpdateAccounts();
-            await db.SaveChangesAsync();
-        }
-
-        [ComponentCommand]
-        public static async Task CraftAccountButton(SocketMessageComponent component, DiscordSocketClient _client, Words _words, IServiceProvider _provider, [ComponentData] string data, ApplicationDbContext db) {
-            //await component.UpdateAsync(x => { x.Content = "Working..."; x.Components = null; });
-            var user = await db.DBUsers.FirstAsync(x => x.DiscordId == component.User.Id);
-            if(user is null) return;
-            var dataObjs = data.Split("|");
-            var account = user.EggIncAccounts.FirstOrDefault(x => x.Id == dataObjs[0]);
-            var quality = (TierInput)int.Parse(dataObjs[1]);
-            var quantity = int.Parse(dataObjs[2]);
-            var requestedArtifact = EggIncArtifacts.GetEiAfxData().artifact_families.FirstOrDefault(x => x.id == dataObjs[3]);
-
-            var contentString = await CraftStringBuilder(account, quantity, quality, requestedArtifact);
-            await component.UpdateAsync(x => { x.Content = contentString; x.Components = null; });
-        }
-
-        private async static Task<string> CraftStringBuilder(EggIncAccount account, int quantity, TierInput quality, ArtifactFamily requestedArtifact) {
-            var stringBuilder = new StringBuilder();
-            var backup = account.Backup;
-            if(backup == null) {
-                return null;
-            }
-
-            backup = new CustomBackup((await ContractsAPI.FirstContact(account.Id)).Backup);
-            stringBuilder.Append($"For **{(string.IsNullOrWhiteSpace(backup.UserName) ? $"Blank account with {backup.EarningsBonus.ToEggString()} EB" : backup.UserName)}** to craft {quantity} T{(int)quality} {requestedArtifact.id}:");
-            stringBuilder.AppendLine();
-
-            var crafter = new Crafter(backup.ArtifactHall);
-            var basket = crafter.GetCraft(quantity, (int)quality, requestedArtifact.id);
-
-            stringBuilder.AppendFormat($"```{"Name",-20}{"Using",-8}{"Need",-8}{"Cost",-8}");
-            stringBuilder.AppendLine();
-            stringBuilder.Append("―――――――――――――――――――――――――――――――――――――――――――");
-            stringBuilder.AppendLine();
-
-            var ingredients = from kvp in basket.GetIngredients()
-                              orderby EggIncArtifacts.GetFamilyShorthand(kvp.Value.Tier.family) ascending, kvp.Value.Tier.tier_number descending
-                              select kvp;
-            foreach(var ingredient in ingredients) {
-                stringBuilder.AppendFormat($"{$"T{ingredient.Value.Tier.tier_number} {EggIncArtifacts.GetFamilyShorthand(ingredient.Value.Tier.family)}",-20}");
-                stringBuilder.AppendFormat($"{ingredient.Value.Use.Format(),-8}");
-                stringBuilder.AppendFormat($"{ingredient.Value.GetNeed(),-8}");
-                stringBuilder.AppendFormat($"{ingredient.Value.Cost.Format(),-8}");
-                stringBuilder.AppendLine();
-            }
-
-            stringBuilder.AppendLine("```");
-            stringBuilder.Append($"Total Cost: **{basket.GetTotalCost().ToString("#,0", new CultureInfo("en-US"))} GE**");
-            stringBuilder.AppendLine();
-            var goldenEggs = backup.GoldenEggsEarned - backup.GoldenEggsSpent;
-            stringBuilder.Append(goldenEggs >= basket.GetTotalCost() ? "You have enough GE!" : "You do not have enough GE!");
-            return stringBuilder.ToString();
-        }
-
         [SlashCommand(Description = "Track your EB since the last time you ran this command")]
         public static async Task TrackEB(FauxCommand command, ApplicationDbContext db, ILogger logger) {
             await command.RespondAsync("Getting backups...");
@@ -438,154 +348,6 @@ Last Backup <t:{backup.LastBackupTime}:R>
                     await command.Channel.SendMessageAsync($"Private callstaff sent. (DMs are blocked)");
                 }
             }
-        }
-
-
-        public enum MERChoice {
-            [Discord.Interactions.ChoiceDisplay("Current")] Current = 0,
-            [Discord.Interactions.ChoiceDisplay("30")] Twenty = 30,
-            [Discord.Interactions.ChoiceDisplay("40")] Thirty = 40,
-            [Discord.Interactions.ChoiceDisplay("50")] Forty = 50
-        };
-
-        [SlashCommand(Description = "Calculate your Mystical Egg Ratio (MER)", ParentCommand = "formulae")]
-        public static async Task Mer(FauxCommand command, ApplicationDbContext db, [SlashParam(Required = true)] MERChoice MERValue) {
-            await command.RespondAsync("Getting account backups...");
-            var user = await db.DBUsers.FirstOrDefaultAsync(x => x.DiscordId == command.User.Id);
-            if(user == null) {
-                await command.ModifyOriginalResponseAsync("⚠️ERROR: Unable to find backups for this user");
-                return;
-            }
-
-            var contentString = "";
-
-            if(user.EggIncAccounts.Count == 1) {
-                contentString = await MERCalculate(user.EggIncAccounts.FirstOrDefault(), user.DiscordUsername, (int)MERValue);
-                await command.ModifyOriginalResponseAsync(x => {
-                    x.Embed = null;
-                    x.Content = $"\n{contentString}\n";
-                });
-            } else {
-                var builder = new ComponentBuilder();
-                foreach(var account in user.EggIncAccounts) {
-                    builder.WithButton($"{account.Name} {account.Backup?.EarningsBonus.ToEggString()}", customId: $"MERAccountButton:{account.Id}|{user.DiscordUsername} - {account.Name}|{(int)MERValue}");
-                }
-                await command.ModifyOriginalResponseAsync(x => { x.Content = "Please select the account you would like to check the MER of."; x.Components = builder.Build(); });
-            }
-        }
-
-        [ComponentCommand]
-        public static async Task MERAccountButton(SocketMessageComponent component, DiscordSocketClient _client, Words _words, IServiceProvider _provider, [ComponentData] string data, ApplicationDbContext db) {
-            var user = await db.DBUsers.FirstAsync(x => x.DiscordId == component.User.Id);
-            if(user is null) return;
-            var dataObjs = data.Split("|");
-            var account = user.EggIncAccounts.FirstOrDefault(x => x.Id == dataObjs[0]);
-            var discordUsername = dataObjs[1];
-            var MERValue = dataObjs[1] == "Current" ? 0 : int.Parse(dataObjs[1]);
-
-            var contentString = await MERCalculate(account, discordUsername, MERValue);
-            await component.UpdateAsync(x => { x.Content = contentString; x.Components = null; });
-        }
-
-        private static async Task<string> MERCalculate(EggIncAccount account, string userName, int MERValue) {
-            double calculateMER(double se, long pe) {
-                var result = (91 * (Math.Log10(se)) + 200 - pe) / 10;
-                return result;
-            }
-
-            string calculateNeededSE(long MER, double se, long pe) {
-                var result = Math.Pow(10, ((10 * MER - 200 + pe) / 91.0)) * 1e18;
-                result -= se;
-                return result.ToEggString();
-            }
-
-            double seQ;
-            double calculateNeededPE(long MER, double se, long pe) {
-                var result = (-10 * MER) + (91 * Math.Log10(seQ)) + 200;
-                result -= pe;
-                return result;
-            }
-
-            var backup = account.Backup;
-            if(backup is null) {
-                return $"Backup not found for user `{account.Name}`";
-            }
-            var seStr = backup.SoulEggs.ToEggString();
-            seQ = backup.SoulEggs / 1e18; // Convert to quintillions
-            var seTotal = backup.SoulEggs;
-            long pe = backup.EggsOfProphecy;
-            var MER = Math.Round(calculateMER(seQ, pe), 2);
-
-            long MERgoal;
-            if(MERValue != 0) {
-                MERgoal = MERValue;
-            } else {
-                double value = Math.Round(calculateMER(seQ, pe), 1);
-                if(value < 30) {
-                    MERgoal = 30;
-                } else if(value < 40) {
-                    MERgoal = 40;
-                } else {
-                    MERgoal = 50;
-                }
-            }
-
-            if(MERgoal > MER) {
-                var MERse = calculateNeededSE(MERgoal, seTotal, pe);
-                return ($"The **MER** for **{userName}** is `{MER}` (<:Egg_of_Prophecy_PE:669981330477547580>`{pe}` and<:Soul_Egg_SE:724341890794913964>`{seStr}`)\nAn additional <:Soul_Egg_SE:724341890794913964>`{MERse}` is needed for MER {MERgoal}");
-            } else {
-                var MERpe = Math.Round(calculateNeededPE(MERgoal, seQ, pe), 1);
-                return ($"The **MER** for **{userName}** is `{MER}` (<:Egg_of_Prophecy_PE:669981330477547580>`{pe}` and<:Soul_Egg_SE:724341890794913964>`{seStr}`)\nYou're able to maintain MER {MERgoal} for another <:Egg_of_Prophecy_PE:669981330477547580>`{MERpe}`");
-            }
-        }
-
-        [SlashCommand(Description = "Calculate your Legendary Luck Coefficient (LLC)", ParentCommand = "formulae")]
-        public static async Task Llc(FauxCommand command, ApplicationDbContext db) {
-            await command.RespondAsync("Getting account backups...");
-            var user = await db.DBUsers.FirstOrDefaultAsync(x => x.DiscordId == command.User.Id);
-            if(user == null) {
-                await command.ModifyOriginalResponseAsync("⚠️ERROR: Unable to find backups for this user");
-                return;
-            }
-
-            var contentString = "";
-
-            if(user.EggIncAccounts.Count == 1) {
-                contentString = await LLCCalculate(user.EggIncAccounts.FirstOrDefault(), user.DiscordUsername);
-                await command.ModifyOriginalResponseAsync(x => {
-                    x.Embed = null;
-                    x.Content = $"\n{contentString}\n";
-                });
-            } else {
-                var builder = new ComponentBuilder();
-                foreach(var account in user.EggIncAccounts) {
-                    builder.WithButton($"{account.Name} {account.Backup?.EarningsBonus.ToEggString()}", customId: $"LLCAccountButton:{account.Id}|{user.DiscordUsername} - {account.Name}");
-                }
-                await command.ModifyOriginalResponseAsync(x => { x.Content = "Please select the account you would like to check the LLC of."; x.Components = builder.Build(); });
-            }
-        }
-
-        [ComponentCommand]
-        public static async Task LLCAccountButton(SocketMessageComponent component, DiscordSocketClient _client, Words _words, IServiceProvider _provider, [ComponentData] string data, ApplicationDbContext db) {
-            var user = await db.DBUsers.FirstAsync(x => x.DiscordId == component.User.Id);
-            if(user is null) return;
-            var dataObjs = data.Split("|");
-            var account = user.EggIncAccounts.FirstOrDefault(x => x.Id == dataObjs[0]);
-            var discordUsername = dataObjs[1];
-
-            var contentString = await LLCCalculate(account, discordUsername);
-            await component.UpdateAsync(x => { x.Content = contentString; x.Components = null; });
-        }
-        private static async Task<string> LLCCalculate(EggIncAccount account, string userName) {
-            var stringBuilder = new StringBuilder();
-            var backup = account.Backup;
-            if(backup == null) {
-                return null;
-            }
-
-            stringBuilder.Append($"The **LLC** for **{userName}** is ``");
-            stringBuilder.AppendLine();
-            return stringBuilder.ToString();
         }
     }
 }
